Fix: subscription plans routing.
